Make dask.compatibility private
